+++ Server/Controller/userController.js
@@ -171,6 +171,51 @@ class UserController {
       });
     }
   }
+
+  static login(req, res) {
+    const { body } = req;
+    const userDetails = {
+      email: body.email,
+      password: body.password,
+    };
+    try {
+      // check if email already exists
+      const emailExists = findUserByEmail(users);
+      if (!emailExists.includes(userDetails.email)) {
+        return res.status(409).json({
+          status: 401,
+          error: 'Authentication Failed',
+          message: 'Request denied',
+        });
+      }
+      const { email, password } = userDetails;
+      const userInfo = users.find(user => email === user.email);
+      const storedPassword = userInfo.password;
+      if (password === storedPassword) {
+        const token = createToken(userInfo);
+        return res.status(201).json({
+          status: 201,
+          data: {
+            token,
+            id: userInfo.id,
+            firstName: userInfo.firstName,
+            lastName: userInfo.lastName,
+            email: userInfo.email,
+          },
+        });
+      }
+      return res.json({
+        status: 401,
+        error: 'Authentication Failed. Incorrect Password',
+        message: 'Request denied',
+      });
+    } catch (e) {
+      return res.status(500).json({
+        status: 500,
+        error: 'Sorry, something went wrong, try again',
+      });
+    }
+  }
 }
 
 export default UserController;
